Now props() callbacks get passed {fnodeForElement: fnodeForElement()} as their second param.
They can use this to look around and do things based on the types, scores, and notes of elements they find. If this turns out to be useful, we'll extend it to score() and note() callbacks.

I don't like that my test has to call rootElement(). Maybe we should pass the document in as well.